[WiP] Range, robot and person visualization
Kept working on the graphical side of the plugin, with range
visualization. Imported icons for robot and generic person.
Currently, when running the application, it is possible to
visualize an example scenario with the robot having a person
in front of it. It is possible to play with the amplitude of the
FOV and Attention cone, even though it is yet to define the role
of these elements