[WiP] plugin resizing, new visual features
Defined a minimum size for the plugin. Defined rules for resizing
with unlimited horizontal extention for the radar.
Extended the GUI to set fov/attention range and decide whether
or not visualing some human-related aspects: distance from
robot, speed (as an arrow), engagement (tbd the type how to
visualize it) and IDs.